Move iterator impls to a new module
